% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_Lifemap.R
\name{add_lm_branches}
\alias{add_lm_branches}
\title{Compute the aes for a subtree.}
\usage{
add_lm_branches(
  proxy,
  aes,
  df,
  df_visible,
  df_descendants,
  group_info,
  all_taxids
)
}
\arguments{
\item{proxy}{The map to be modified.}

\item{aes}{The dataframe containing the aesthetics informations (must be of lm_branches class).}

\item{df}{The full dataframe.}

\item{df_visible}{The dataframe containing visible taxa.}

\item{df_descendants}{The dataframe containing all the information on the descendants of visible taxa.}

\item{group_info}{The points' group.}

\item{all_taxids}{A vector with all the visible taxids and their direct descendants.}
}
\value{
A list of values.
}
\description{
Compute the aes for a subtree.
}
