% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_Lifemap.R
\name{add_lm_markers}
\alias{add_lm_markers}
\title{Compute the aes for a set of points.}
\usage{
add_lm_markers(proxy, aes, df, df_visible, group_info)
}
\arguments{
\item{proxy}{The map to be modified.}

\item{aes}{The dataframe containing the aesthetics informations (must be of lm_markers class).}

\item{df}{The full dataframe.}

\item{df_visible}{The dataframe containing visible taxa.}

\item{group_info}{The points' group / layerID.}
}
\value{
A list of values.
}
\description{
Compute the aes for a set of points.
}
