% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_Lifemap.R
\name{add_lm_piecharts}
\alias{add_lm_piecharts}
\title{Compute the aes for discret values.}
\usage{
add_lm_piecharts(proxy, aes, df, df_visible, layer)
}
\arguments{
\item{proxy}{The map to be modified.}

\item{aes}{The dataframe containing the aesthetics informations (must be of lm_piecharts class).}

\item{df}{The full dataframe.}

\item{df_visible}{The dataframe containing visible taxa.}

\item{layer}{The points' group.}
}
\value{
A list of values.
}
\description{
Compute the aes for discret values.
}
