% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Lifemap.R
\name{build_Lifemap}
\alias{build_Lifemap}
\title{A function to construct a LifemapR object, usable by other functions}
\usage{
build_Lifemap(df, basemap = c("fr", "ncbi", "base", "virus"), verbose = TRUE)
}
\arguments{
\item{df}{a dataframe containing at least a column named "taxid"}

\item{basemap}{the basemap wanted ("fr","ncbi", "base" or "virus")}

\item{verbose}{if TRUE, will write informations in the terminal}
}
\value{
a lifemap object with :
\itemize{
\item df : a dataframe containing at least :
\itemize{
\item the taxids (taxid)
\item the longitude (long)
\item the latitude (lat)
\item the scientific nam (sci_name)
\item the zoom at which the taxa can be seen (zoom)
\item the ascendants of requested taxids (ascend)
\item the type of each taxid ("requested" or "ancestor")
}
\item basemap : the basemap used to get taxa's informations
}
}
\description{
A function to construct a LifemapR object, usable by other functions
}
\examples{
data(eukaryotes_1000)
LM <- build_Lifemap(eukaryotes_1000, "fr")
}
