% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrix.R
\name{create_matrix}
\alias{create_matrix}
\title{Create a dataframe for the ancestry}
\usage{
create_matrix(df, cols)
}
\arguments{
\item{df}{a dataframe containing taxids and values}

\item{cols}{the columns that contains the values}
}
\value{
a dataframe
}
\description{
Create a dataframe with couples of taxids, each taxid is associated with each
one of it's ancestors, values are also associated with the taxids
}
\examples{
data(LM_eukaryotes)

create_matrix(LM_eukaryotes$df, c("GC.", "Genes"))
}
