% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{display_map}
\alias{display_map}
\title{Create a Lifemap base}
\usage{
display_map(df = NULL, map = "ncbi")
}
\arguments{
\item{map}{The base choose to display, either \itemize{
\itemize{
\item 'fr' for \url{https://lifemap-fr.univ-lyon1.fr/}
\item 'ncbi' for \url{https://lifemap-ncbi.univ-lyon1.fr/} }
}}
}
\value{
HTML widget object  with graphics layers
}
\description{
This function create a blank Leaflet map using the leaflet function from the leaflet package.
if a dataframe is provided, it will be used for the creation of the map object
}
\examples{
display_map("fr")
}
