% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_Lifemap.R
\name{display_option}
\alias{display_option}
\title{compute the different display options}
\usage{
display_option(m, aes, df, type, leaves, i)
}
\arguments{
\item{m}{a map}

\item{aes}{the aesthetics}

\item{df}{a dataframe containing all the taxids}

\item{type}{c("markers", "discret")}

\item{leaves}{vector of all terminal taxids}

\item{i}{index of the aesthetics}
}
\value{
a map
}
\description{
compute the different display options
}
