% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_Lifemap.R
\name{draw_Lifemap}
\alias{draw_Lifemap}
\title{Represent continuous datas on a Lifemap background}
\usage{
draw_Lifemap(lm_obj)
}
\arguments{
\item{lm_obj}{a Lifemap object filled with aesthetics}
}
\value{
a shiny application
}
\description{
draw a map and all the aesthetics in the order you put them, the last one will be on top of the others
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
data(LM_eukaryotes)
lifemap(LM_eukaryotes) + lm_markers() + lm_branches()
}
}
