% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gen_res}
\alias{gen_res}
\title{Genomic results}
\format{
A data frame with 808 rows and 3 variables:
\describe{
\item{taxid}{NCBI taxid}
\item{Genome_size}{the Genome size in pb}
\item{TEcontent_bp}{the transposable element content in pb}
}
}
\usage{
data(gen_res)
}
\description{
A dataset containing information on Genome size and TE content for 808 taxids
}
\keyword{datasets}
