% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_coordinates}
\alias{get_coordinates}
\title{Get coordinates from SolR database}
\usage{
get_coordinates(list_IDs)
}
\arguments{
\item{list_IDs}{a vector of TaxIDs you want to represent}
}
\value{
a dataframe
}
\description{
This function return a dataframe with the coordinates on the lifemap for each ID
}
\examples{
get_coordinates(c(2,9443,2087))
}
