% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Lifemap.R
\name{get_direct_ancestor}
\alias{get_direct_ancestor}
\title{Add the direct ancestor for each taxa of the dataframe.}
\usage{
get_direct_ancestor(df)
}
\arguments{
\item{df}{A dataframe with requested taxid and the full ascendance information.}
}
\value{
A dataframe with the direct ancestor for each taxa.
}
\description{
Add the direct ancestor for each taxa of the dataframe.
}
