% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifemap_obj.R
\name{is.lifemap_obj}
\alias{is.lifemap_obj}
\title{Reports whether x is a lm_obj object}
\usage{
is.lifemap_obj(x)
}
\arguments{
\item{x}{the object to test}
}
\description{
Reports whether x is a lm_obj object
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
data(LM_eukaryotes)
is.lifemap_obj(LM_eukaryotes)
}

}
