% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_markers.R
\name{is.lm_markers}
\alias{is.lm_markers}
\title{Reports whether x is a lm_markers object}
\usage{
is.lm_markers(x)
}
\arguments{
\item{x}{the object to test}
}
\description{
Reports whether x is a lm_markers object
}
