% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_piecharts.R
\name{is.lm_piecharts}
\alias{is.lm_piecharts}
\title{Reports whether x is a lm_branches object.}
\usage{
is.lm_piecharts(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A boolean indicating whether or not the object is of lm_piecharts type.
}
\description{
Reports whether x is a lm_branches object.
}
