% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kraken_res}
\alias{kraken_res}
\title{Kraken results}
\format{
A data frame with 4427 rows and 6 variables:
\describe{
\item{coverage_percent}{Percentage of fragments covered by the clade rooted at this taxon}
\item{coverage_number}{Number of fragments covered by the clade rooted at this taxon}
\item{fragment_number}{Number of fragments assigned directly to this taxon}
\item{rank}{A rank code, indicating (U)nclassified, (R)oot, (D)omain, (K)ingdom, (P)hylum, (C)lass, (O)rder, (F)amily, (G)enus, or (S)pecies. Taxa that are not at any of these 10 ranks have a rank code that is formed by using the rank code of the closest ancestor rank with a number indicating the distance from that rank. E.g., "G2" is a rank code indicating a taxon is between genus and species and the grandparent taxon is at the genus rank.}
\item{taxid}{NCBI taxonomic ID number}
\item{name}{Indented scientific name}
}
}
\usage{
data(kraken_res)
}
\description{
A dataset containing NCBI information of 1000 eukaryotes. The variables are as follows:
}
\keyword{datasets}
