% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifemap.R
\name{lifemap}
\alias{lifemap}
\title{create a new Lifemap visualisation}
\usage{
lifemap(lm_obj, zoom = 4)
}
\arguments{
\item{lm_obj}{Lifemap object used for data visualisation}
}
\value{
a lifemap object
}
\description{
initialise a lifemap object. It can be used to describe aesthetics like legend's properties ...
}
