% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_branches.R
\name{lm_branches}
\alias{lm_branches}
\title{add a layer to a Lifemap object}
\usage{
lm_branches(
  data = NULL,
  color = NULL,
  var_color = NULL,
  size = 5,
  min = 2,
  max = 20,
  opacity = 0.5,
  FUN = NULL,
  legend = TRUE,
  legendPosition = c("topright", "bottomright", "bottomleft", "topleft")
)
}
\arguments{
\item{data}{a sub dataset to use, if NULL then all of the taxids from the lifemap object given to lifemap() will be used}

\item{color}{either a color for the branches or a palette if a variable is used to rperesent branches' color}

\item{var_color}{a column name of the original dataframe to represent this variable by the color of branches}

\item{size}{either a numeric for the branche's thickness or a variable to be représented by the branche's thickness}

\item{min}{an integer indicating the minimal thickness of the branches if the size represent a variable}

\item{max}{an integer indicating the maximal thickness of the branches if the size represent a variable}

\item{opacity}{an integer indicating branche's opacity}

\item{FUN}{the function to be applied to infer values}

\item{legend}{a logical indicating whether or not to display the legend}

\item{legendPosition}{c("topright", "bottomright", "bottomleft", "topleft"), the position of the legend}
}
\value{
a lifemap object
}
\description{
add a layer to a Lifemap object
}
\examples{
data(LM_eukaryotes)

lm_branches(var_color = "GC.", color = "Accent")

lm_branches(data = LM_eukaryotes$df[LM_eukaryotes$df$Group \%in\% "Plants",])

}
