% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LM_eukaryotes}
\alias{LM_eukaryotes}
\title{Transformation in a LifemapR format of NCBI information for 1000 eukaryotes}
\format{
A lifemap object - a list containing the basemap used to fetch data and df, a data frame with 2760 rows and 26 variables:
\describe{
\item{X.Organism.Name}{Organism name at the species level}
\item{taxid}{NCBI taxid}
\item{BioProject.Accession}{BioProject Accession number (from BioProject database)}
\item{BioProject.ID}{BioProject ID}
\item{Group}{Commonly used organism groups:  Animals, Fungi, Plants, Protists}
\item{SubGroup}{NCBI Taxonomy level below group: Mammals, Birds, Fishes, Flatworms, Insects, Amphibians, Reptiles, Roundworms, Ascomycetes, Basidiomycetes, Land Plants, Green Algae, Apicomplexans, Kinetoplasts}
\item{Size..Mb.}{Total length of DNA submitted for the project}
\item{GC.}{Percent of nitrogenous bases (guanine or cytosine) in DNA submitted for the project}
\item{Assembly.Accession}{Name of the genome assembly (from NCBI Assembly database)}
\item{Replicons}{Number of replicons in the assembly}
\item{WGS}{Four-letter Accession prefix followed by version as defined in WGS division of GenBank/INSDC}
\item{Scaffolds}{Number of scaffolds in the assembly}
\item{Genes}{Number of Genes annotated in the assembly}
\item{Proteins}{Number of Proteins annotated in the assembly}
\item{Release.Date}{First public sequence release for the project}
\item{Modify.Date}{Sequence modification date for the project}
\item{Status}{Highest level of assembly: \if{html}{\out{<br>}} Chromosomes: one or more chromosomes are assembled\if{html}{\out{<br>}} Scaffolds or contigs: sequence assembled but no chromosomes}
\item{Center}{Origin of the sample}
\item{BioSample.Accession}{BioSample Accession number}
\item{lon}{longitude of taxids on a specific basemap}
\item{lat}{latitude of taxids on a specific basemap}
\item{sci_name}{scientific name of taxids}
\item{zoom}{zoom of taxids on a specific basemap}
\item{ascend}{the list of all ancestors of taxids on a specific basemap}
\item{type}{either "requested" if the taxid was given, "ancestor" if gotten from the database}
\item{ancestor}{the direct ancestor oftaxids on a specific basemap}
}
}
\usage{
data(LM_eukaryotes)
}
\description{
A dataset containing NCBI information of 1000 eukaryotes. The variables are as follows:
}
\keyword{datasets}
