% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_markers.R
\name{lm_markers}
\alias{lm_markers}
\title{add a layer to a Lifemap object}
\usage{
lm_markers(
  data = NULL,
  radius = 10,
  min = 10,
  max = 40,
  fillColor = NULL,
  var_fillColor = NULL,
  fillOpacity = 0.8,
  stroke = FALSE,
  color = NULL,
  var_color = NULL,
  weight = 1,
  opacity = 0.8,
  legend = TRUE,
  legendPosition = c("topright", "bottomright", "bottomleft", "topleft"),
  legendOrientation = c("vertical", "horizontal"),
  legendOpacity = 0.5,
  FUN = NULL,
  display = c("auto", "requested", "all", "leaves"),
  popup = NULL,
  label = NULL
)
}
\arguments{
\item{data}{a sub dataset to use, if NULL then all of the taxids from the lifemap object given to lifemap() will be used}

\item{radius}{either a numerical value or a column name of the original dataframe to represent this variable by the size of points}

\item{min}{an integer indicating the minimal size of the points if radius represent a variable}

\item{max}{an integer indicating the maximal size of the points if radius represent a variable}

\item{fillColor}{either a color or a palette if the fillColor is represented by a variable}

\item{var_fillColor}{a column name of the original dataframe to represent this variable by the fillColor of points}

\item{fillOpacity}{a numeric indicating the fill opacity}

\item{stroke}{a logical indicating whether to draw a border for points}

\item{color}{the stroke color. Either a color or a palette if the stroke color is represented by a variable}

\item{var_color}{a column name of the original dataframe to represent this variable by the stroke color of points}

\item{weight}{stroke width in pixels}

\item{opacity}{stroke opacity}

\item{legend}{whether to display the legend}

\item{legendPosition}{c("topright", "bottomright", "bottomleft", "topleft"),}

\item{legendOrientation}{c("vertical", "horizontal"),}

\item{legendOpacity}{legend's opacity (apply on the shapes in the legend, not the background itself)}

\item{FUN}{the function to be applied to the variables if NULL then the information missing from the parent nodes won't be inferred}

\item{display}{c("auto", "requested", "all", "leaves"), a string indicating how to display points :
\itemize{
\item "auto" : the markers are displayed depending on the zoom, by default, allow to have a lot of points
\item "requested" : only display the requested taxa, but all at the same time
\item "all" : display all the taxa including all the ancestors to the root
\item "leaves" : display only the last (most recent) taxa
}

(WARNING : "requested" and "auto" shouldn't be used to display more than 1000 markers as it may result in long computing time)}

\item{popup}{a column name indicating what to display when clicking on a node}

\item{label}{a column name indiacting what to display when hovering on a node}
}
\value{
a list containing all aesthetics informations for one serie of markers
}
\description{
Allow to add a layer with circles that can represent data. The parameters that can be used to represent data are the following:
\itemize{
\item radius
\item fillColor
\item stroke
\item color
}
}
\examples{
data(LM_eukaryotes)

lm_markers(data = LM_eukaryotes$df[LM_eukaryotes$df$Group \%in\% "Plants", ])

lm_markers(radius = "GC.", var_fillColor = "Genes")

}
