% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_piecharts.R
\name{lm_piecharts}
\alias{lm_piecharts}
\title{add a layer to a Lifemap object}
\usage{
lm_piecharts(
  data = NULL,
  param,
  type = c("pie", "bar", "polar-area", "polar-radius", "auto"),
  width = 30,
  height = 30,
  opacity = 1,
  showLabels = FALSE,
  pal = "Accent",
  legend = TRUE,
  legendPosition = c("topright", "bottomright", "bottomleft", "topleft"),
  display = c("auto", "requested", "all", "leaves")
)
}
\arguments{
\item{data}{a sub dataset to use, if NULL then all of the taxids from the lifemap object given to lifemap() will be used}

\item{param}{a string indicating the discret variable to be represented}

\item{type}{the type of chart to draw}

\item{width}{the maximal width of the charts}

\item{height}{the maximal height of the charts}

\item{opacity}{the chart's opacity}

\item{showLabels}{a boolean indicating whether to display the values directly on the chart or not}

\item{pal}{the palette to be used for the charts}

\item{legend}{whether to draw the legend or not}

\item{legendPosition}{where should the legend be placed}

\item{display}{c("auto", "requested", "all", "leaves"), a string indicating how to display points :
\itemize{
\item "auto" : the markers are displayed depending on the zoom, by default, allow to have a lot of points
\item "requested" : only display the requested taxa, but all at the same time
\item "all" : display all the taxa including all the ancestors to the root
\item "leaves" : display only the last (most recent) taxa
}}
}
\value{
a lifemap object
}
\description{
add a layer to a Lifemap object
}
\examples{
data(LM_eukaryotes)

lm_piecharts(param = "Status")

lm_piecharts(data = LM_eukaryotes$df[LM_eukaryotes$df$Group \%in\% "Plants", ], param = "Status")

}
