% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmpar.R
\name{lmpar}
\alias{lmpar}
\title{lmpar}
\usage{
lmpar(map = "default")
}
\arguments{
\item{map}{a character string for the version of lifemap among 'standard', 'ncbi', 'french' and virus' version; 'standard' by default}
}
\value{
Get or set the name of the basemap.
}
\description{
Get or set the basemap for lifemap as an option.
This function is called by others to know what map to
display and what database to query to get the taxa
coordinates.
}
