% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_newick.R
\name{make_newick}
\alias{make_newick}
\title{Create a newick for the given dataset.}
\usage{
make_newick(df)
}
\arguments{
\item{df}{A dataframe either enriched with by the build_Lifemap function or containing at least two columns :
\itemize{
\item "taxid" containing NCBI format taxids
\item "ancestor" containing for each taxid, it's direct ancestor in the phylogeny
}}
}
\value{
A character string.
}
\description{
Create a newick for the given dataset.
}
