% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrix.R
\name{pass_infos}
\alias{pass_infos}
\title{Infer values to nodes}
\usage{
pass_infos(M, FUN, value)
}
\arguments{
\item{M}{the dataframe returned by create_matrix}

\item{FUN}{the function to be applied when passing the info}

\item{value}{the column to which the function apply}
}
\value{
an array of values
}
\description{
Infer values to nodes
}
\examples{
data(LM_eukaryotes)

infos <- create_matrix(LM_eukaryotes$df, c("GC.", "Genes"))

inferred_values <- pass_infos(M = infos, FUN = mean, value = "GC.")
}
