% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrix.R
\name{pass_infos_discret}
\alias{pass_infos_discret}
\title{Pass discret informations to nodes}
\usage{
pass_infos_discret(M, value)
}
\arguments{
\item{M}{The dataframe returned by create_matrix.}

\item{value}{The column name to which the function applies.}
}
\value{
A dataframe.
}
\description{
Pass discret informations to nodes
}
\examples{
data(LM_eukaryotes)

infos <- create_matrix(LM_eukaryotes$df, "Status")

inferred_values <- pass_infos_discret(M = infos, value = "Status")
}
