% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Lifemap_new.R
\name{request_database_new}
\alias{request_database_new}
\title{Request one of the core of the solr database corresponding to the basemap choosen with TaxIDs wanted}
\usage{
request_database_new(taxids, basemap)
}
\arguments{
\item{taxids}{a vector of TaxIDs to be requested}

\item{basemap}{the name of the basemap wanted to represent the datas on (can be "ncbi", "fr" ...)}
}
\value{
a dataframe containing all the informations requested from the database
}
\description{
Request one of the core of the solr database corresponding to the basemap choosen with TaxIDs wanted
}
