% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxid2map.R
\name{taxid2map}
\alias{taxid2map}
\title{taxid2map}
\usage{
taxid2map(taxid, verbose = TRUE)
}
\arguments{
\item{taxid}{either a unique taxid or a vector of taxids for which to
retrieve coordinates}

\item{verbose}{Boolean. Should warnings be written if case of duplicated or invalid
taxids. Default to TRUE.
retrieve coordinates}
}
\value{
A list with two components:
coo: a dataframe with information for the taxid(s) requested:
lat, long, scientific name, zoom at which the taxa is visible, and
number of descendants of this taxa.
missing: a
}
\description{
This function queries the Lifemap database to retrieve
the coordinates of taxa on Lifemap. Queries are based on ncbi taxid(s).
}
\examples{
taxid2map(c(2,9443,2087))
}
