% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_ancestry.R
\name{trace_ancestry}
\alias{trace_ancestry}
\title{Trace the ancestry of multiple taxa}
\usage{
trace_ancestry(ids, map)
}
\arguments{
\item{ids}{a list of TaxID to visualise}

\item{map}{the base map wanted}
}
\value{
a Leaflet map
}
\description{
Trace the ancestry of multiple taxa
}
\examples{
trace_ancestry(c(3641,3649,403667,3394,2), "fr")
}
