% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_Lifemap.R
\name{add_lm_branches}
\alias{add_lm_branches}
\title{compute the aes for a subtree}
\usage{
add_lm_branches(
  proxy,
  aes,
  df,
  df_visible,
  df_descendants,
  group_info,
  all_taxids
)
}
\arguments{
\item{proxy}{the map to be modified}

\item{aes}{the dataframe containing the aesthetics informations (must be of lm_markers class)}

\item{df}{the full dataframe}

\item{df_visible}{the dataframe containing visible taxa}

\item{df_descendants}{the dataframe containing all the information on the descendants of visible taxa}

\item{group_info}{the points' group}

\item{all_taxids}{a vector with all the visible taxids and their direct descendants}
}
\value{
a list of values
}
\description{
compute the aes for a subtree
}
