% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_Lifemap.R
\name{add_lm_branches}
\alias{add_lm_branches}
\title{Compute the aesthetics for a subtree visualisation.}
\usage{
add_lm_branches(
  proxy,
  aes,
  df,
  df_visible,
  df_descendants,
  group_info,
  all_taxids
)
}
\arguments{
\item{proxy}{The map to be modified.}

\item{aes}{The dataframe containing the aesthetics details (must be of lm_branches class).}

\item{df}{The full dataframe.}

\item{df_visible}{The dataframe containing visible taxa.}

\item{df_descendants}{The dataframe containing all the information on the descendants of visible taxa.}

\item{group_info}{the ID of this group of lines.}

\item{all_taxids}{A vector containing all the visible taxids and their direct descendants.}
}
\value{
An updated map with the new layer added.
}
\description{
Compute the aesthetics for a subtree visualisation.
}
