% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_Lifemap.R
\name{add_lm_markers}
\alias{add_lm_markers}
\title{compute the aes for a set of points}
\usage{
add_lm_markers(proxy, aes, df, df_visible, group_info)
}
\arguments{
\item{proxy}{the map to be modified}

\item{aes}{the dataframe containing the aesthetics informations (must be of lm_markers class)}

\item{df}{the full dataframe}

\item{df_visible}{the dataframe containing visible taxa}

\item{group_info}{the points' group}
}
\value{
a list of values
}
\description{
compute the aes for a set of points
}
