% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Lifemap.R
\name{build_Lifemap}
\alias{build_Lifemap}
\title{A function to construct a LifemapR object, usable by the other functions of the package.}
\usage{
build_Lifemap(df, basemap = c("fr", "ncbi", "base", "virus"), verbose = TRUE)
}
\arguments{
\item{df}{A dataframe containing at least a column named "taxid" containing NCBI format TaxIDs and that can contain characteristics associated with those TaxIDs.}

\item{basemap}{The basemap wanted ("fr","ncbi", "base" or "virus").}

\item{verbose}{If TRUE, will write informations on the status of the operation in the terminal.}
}
\value{
A lifemap object with :
\itemize{
\item df : a dataframe containing at least :
\itemize{
\item The taxids (taxid)
\item The longitude (long)
\item The latitude (lat)
\item The scientific nam (sci_name)
\item The zoom at which the taxa can be seen (zoom)
\item The ascendants of requested taxids (ascend)
\item The type of each taxid ("requested" or "ancestor")
}
\item basemap : the basemap used to get taxa's informations
}
}
\description{
A function to construct a LifemapR object, usable by the other functions of the package.
}
\examples{
data(eukaryotes_1000)
LM <- build_Lifemap(eukaryotes_1000, "fr")
}
