% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrix.R
\name{create_matrix}
\alias{create_matrix}
\title{Create a dataframe for the ancestry.}
\usage{
create_matrix(df, cols)
}
\arguments{
\item{df}{A dataframe containing taxids and values.}

\item{cols}{The columns containing the values which needs to be inferred.}
}
\value{
A dataframe.
}
\description{
Create a dataframe with pairs of taxids, each taxid and their corresponding values are associated with every one of its ancestors.
}
\examples{
data(LM_eukaryotes)

create_matrix(LM_eukaryotes$df, c("GC.", "Genes"))
}
