% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_Lifemap.R
\name{create_value_range}
\alias{create_value_range}
\title{Compute a new scale for a value}
\usage{
create_value_range(value, df, df2, min, max)
}
\arguments{
\item{value}{A vector of values.}

\item{df}{The full dataframe.}

\item{df2}{The dataframe containing visibles taxa.}

\item{min}{The new minimum of the range.}

\item{max}{The new maximum of the range.}
}
\value{
A vector of values.
}
\description{
Compute a new scale for a value
}
