% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_map.R
\name{display_map}
\alias{display_map}
\title{Create a Lifemap base.}
\usage{
display_map(df = NULL, basemap = c("fr", "ncbi", "base", "virus"))
}
\arguments{
\item{df}{A dataframe. If given, its columns can be easily accessed with "~" (eg. ~GC.).}

\item{basemap}{The basemap choosen to be displayed, it can be either :
\itemize{
\item 'fr' for \url{https://lifemap-fr.univ-lyon1.fr/}
\item 'ncbi' for \url{https://lifemap-ncbi.univ-lyon1.fr/}
\item 'base' for \url{https://lifemap.univ-lyon1.fr/}
\item 'virus' for \url{https://virusmap.univ-lyon1.fr/}
}}
}
\value{
An HTML widget object  with graphics layers.
}
\description{
This function create a blank Leaflet map using the leaflet function from the leaflet package.
If a dataframe is provided, it will be used for the creation of the map.
}
\examples{
display_map()
}
