% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_Lifemap.R
\name{display_option}
\alias{display_option}
\title{Compute the different display options.}
\usage{
display_option(m, aes, df, type, leaves, i)
}
\arguments{
\item{m}{The map to be modified.}

\item{aes}{The dataframe containing the aesthetics details}

\item{df}{The full dataframe.}

\item{type}{A string indicating the type of representation, either "markers" or "discret"}

\item{leaves}{The Vector of all the terminal taxids.}

\item{i}{The index of the aesthetics.}
}
\value{
An updated map.
}
\description{
Compute the different display options.
}
