% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_Lifemap.R
\name{draw_Lifemap}
\alias{draw_Lifemap}
\title{Represent data on a Lifemap basemap.}
\usage{
draw_Lifemap(lm_obj)
}
\arguments{
\item{lm_obj}{A Lifemap object filled with aesthetics.}
}
\value{
A shiny application
}
\description{
Draw a map and all the aesthetics in the order you put them in, the last one will be on top of the others.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
  data(LM_eukaryotes)
  lifemap(LM_eukaryotes) + lm_markers() + lm_branches()
}
}
