% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geturlprefix.R
\name{geturlprefix}
\alias{geturlprefix}
\title{geturlprefix}
\usage{
geturlprefix(basemap = NULL)
}
\arguments{
\item{basemap}{To which basemap are the data retrieved associated.
By default, the basemap retrieved by function \code{lmpar()} is used.}
}
\value{
Get or set the name of the basemap.
}
\description{
Gets the prefix of the url used for accessing Lifemap
based on the basemap option set with lmpar().
This is mainly used as an internal function.
}
