% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifemap_obj.R
\name{is.lifemap_obj}
\alias{is.lifemap_obj}
\title{Reports whether x is a lifemap_obj object.}
\usage{
is.lifemap_obj(x)
}
\arguments{
\item{x}{The object to test.}
}
\value{
A boolean indicating whether or not the object is of lifemap_obj type.
}
\description{
Reports whether x is a lifemap_obj object.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
data(LM_eukaryotes)
is.lifemap_obj(LM_eukaryotes)
}

}
