% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifemap.R
\name{lifemap}
\alias{lifemap}
\title{lifemap}
\usage{
lifemap(data = NULL, ...)
}
\arguments{
\item{data}{a data frame with at least two columns (lat and long). This
data frame can be (but does not have to) the result of a call to \code{taxid2map()}.}

\item{...}{other options, as passed to \code{leaflet()}}
}
\value{
A map widget.
}
\description{
Create a new lifemap using the basemap set by \code{lmpar()}.
}
