% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifemap.R
\name{lifemap}
\alias{lifemap}
\title{create a new Lifemap visualisation}
\usage{
lifemap(lm_obj, zoom = 4)
}
\arguments{
\item{lm_obj}{Lifemap object used for data visualisation}

\item{zoom}{the level of zoom for which values are displayed (if zoom = 0 only the nodes at the current zoom level or lower can have their values displayed)}
}
\value{
a lifemap object
}
\description{
Initialise a lifemap object.
It can be used to describe aesthetics like the zoom level at which taxids becomes visible...
}
