% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifemap.R
\name{lifemap}
\alias{lifemap}
\title{Initialise a new Lifemap visualisation.}
\usage{
lifemap(lm_obj, zoom = 4)
}
\arguments{
\item{lm_obj}{lifemap_obj object used for data visualisation.}

\item{zoom}{The level of zoom for which values are displayed (if zoom = 0 only the nodes at the current zoom level or lower can have their values displayed).}
}
\value{
A lifemap_obj object.
}
\description{
Initialise a lifemap_obj object.
It can be used to describe aesthetics like the zoom level at which taxids becomes visible...
}
