% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrix.R
\name{pass_infos}
\alias{pass_infos}
\title{Infer numerical values to nodes.}
\usage{
pass_infos(M, FUN, value)
}
\arguments{
\item{M}{The dataframe returned by create_matrix.}

\item{FUN}{The function to be applied when inferring the values.}

\item{value}{The column name to which the function applies.}
}
\value{
An array of values.
}
\description{
Infer numerical values to nodes.
}
\examples{
data(LM_eukaryotes)

infos <- create_matrix(LM_eukaryotes$df, c("GC.", "Genes"))

inferred_values <- pass_infos(M = infos, FUN = mean, value = "GC.")
}
