% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_matrix.R
\name{pass_infos_discret}
\alias{pass_infos_discret}
\title{Infer discret values to nodes for lm_piecharts function.}
\usage{
pass_infos_discret(M, value)
}
\arguments{
\item{M}{The dataframe returned by create_matrix.}

\item{value}{The column name represented as piecharts.}
}
\value{
A dataframe containing the TaxIDs and as many columns as there are distinct values.
}
\description{
Infer discret values to nodes for lm_piecharts function.
}
\examples{
data(LM_eukaryotes)

infos <- create_matrix(LM_eukaryotes$df, "Status")

inferred_values <- pass_infos_discret(M = infos, value = "Status")
}
