% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifemap_obj.R
\name{print.lifemap_obj}
\alias{print.lifemap_obj}
\title{Method to print lifemap_obj objects.}
\usage{
\method{print}{lifemap_obj}(x, ...)
}
\arguments{
\item{x}{An lifemap_obj.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Either a description of the dataframe and basemap used for the lm_obj object, or a shiny application if aesthetics are furnished
}
\description{
Method to print lifemap_obj objects.
}
\examples{
## Only run examples in interactive R sessions
if (interactive()) {
data(LM_eukaryotes)
print(LM_eukaryotes)
}
}
