% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxid2map.R
\name{taxid2map}
\alias{taxid2map}
\title{taxid2map}
\usage{
taxid2map(
  taxid,
  remove.missing = TRUE,
  remove.duplicates = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{taxid}{either a unique taxid or a vector of taxids for which to
retrieve coordinates}

\item{remove.missing}{Logical. Should taxids that were not found
be removed from the output. Default to TRUE.}

\item{remove.duplicates}{Logical. Should duplicated taxids (if any)
be removed from the output. Default to TRUE.}

\item{verbose}{Logical. Should warnings be written if case of duplicated or invalid
taxids. Default to TRUE.
retrieve coordinates}
}
\value{
A data frame with lon, lat and zoom, with the first column
in the same order as the input taxid.
}
\description{
This function queries the Lifemap database to retrieve
the coordinates of taxa on Lifemap. Queries are based on ncbi taxid(s).
}
\examples{
taxid2map(c(2,9443,2087))
}
