% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Lifemap.R, R/build_Lifemap_new.R
\name{url_verification}
\alias{url_verification}
\title{Check if the URL is working}
\usage{
url_verification(basemap_url, t = 20)

url_verification(basemap_url, t = 20)
}
\arguments{
\item{basemap_url}{the url corresponding to the basemap interrogated}

\item{t}{the time before timeout}
}
\value{
a logical

a logical
}
\description{
Check if the URL is working

Check if the URL is working
}
