% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_Lifemap.R
\name{url_verification}
\alias{url_verification}
\title{Check if the URL is working}
\usage{
url_verification(basemap_url, t = 20)
}
\arguments{
\item{basemap_url}{The url corresponding to the basemap interrogated.}

\item{t}{The time before timeout.}
}
\value{
A logical.
}
\description{
Check if the URL is working
}
